LayoutTests/imported/w3c/web-platform-tests/webrtc/protocol/h264-profile-levels.https_rest-expected.txt


FAIL Level 1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 99 but got 14400
FAIL Level 2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 396 but got 14400
FAIL Level 3 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 1620 but got 14400
NOTRUN Level 6 H264 video is appropriately constrained Level 6 H264 video is not supported
FAIL Level 1.1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 396 but got 14400
FAIL Level 1.2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 396 but got 14400
FAIL Level 1.3 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 396 but got 14400
FAIL Level 2.1 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 792 but got 14400
FAIL Level 2.2 H264 video is appropriately constrained assert_less_than_equal: frame size expected a number less than or equal to 1620 but got 14400
NOTRUN Level 6.1 H264 video is appropriately constrained Level 6.1 H264 video is not supported
NOTRUN Level 6.2 H264 video is appropriately constrained Level 6.2 H264 video is not supported
